requirements.txt
# HTP 心理分析系统依赖

# Streamlit 核心
streamlit>=1.28.0

# LangChain 相关
langchain-core>=0.1.0
langchain-aws>=0.1.0
langgraph>=0.0.40

# AWS Bedrock
boto3>=1.34.0

# 图像处理
Pillow>=10.0.0

# 数据处理
PyYAML>=6.0
python-dotenv>=1.0.0

# 可选组件 (如果使用绘画功能)
streamlit-tldraw>=0.1.0  # 需要单独安装

# 开发依赖 (可选)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
